backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pinecone-client
openai
transformers
ollama
stability-sdk